pyproject.toml
[project]
name = "LeafDoctor"
version = "1.0.0"
description = "A web that helps users classify leaf diseases at ease"
dependencies = [
    "flask",
    "flask-cors",
    "tensorflow",
    "numpy",
    "gdown",
    "cloudinary",
    "python-dotenv",
]

[build-system]
requires = ["flit_core<4"]
build-backend = "flit_core.buildapi"